Hardware/BLE/Clients: Added base class BaseAppleMedia
